[Fixed] Sprite、SpriteManagerクラスの実装とそれに伴う変更

+++ Project2/Actor.h
@@ -17,6 +17,6 @@ public:
 
 protected:
 	std::shared_ptr<got::Texture> texture;
-	got::Vector2<float> position;
+	got::Vector2<int> position;
 
 };
